03-15 15:56 팀 프로젝트 개인 수정

+++ src/main/java/com/green/hoteldog/board/models/BoardCommentVo.java
@@ -8,9 +8,9 @@ import java.util.List;
 @Data
 public class BoardCommentVo {
     @Schema(description = "최대 페이지 수")
-    private int commentMaxPage;
+    private long commentMaxPage;
     @Schema(description = "댓글 개수")
-    private int commentCount;
+    private long commentCount;
     @Schema(description = "댓글 정보")
     private List<CommentInfoVo> commentInfoVoList;
 }
