+++ src/main/java/com/green/hoteldog/board/models/GetBoardInfoDto.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class GetBoardInfoDto {
     @Min(value = 1,message = "boardPk 값은 1 이상이어야 합니다.")
     @Schema(description = "게시글pk",defaultValue = "1")
-    private int boardPk;
+    private long boardPk;
 
 }
